% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{guess_locale}
\alias{guess_locale}
\title{Encoding guess}
\usage{
guess_locale(.file)
}
\arguments{
\item{.file}{A path to a file or a connection.}
}
\value{
A \code{\link[readr]{locale}} object.
}
\description{
Guess the locale for \code{.file} via \code{\link[readr]{guess_encoding}}.
}
