% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{locate_pattern}
\alias{locate_pattern}
\title{Locate index of pattern}
\usage{
locate_pattern(.lines, .pattern)
}
\arguments{
\item{.lines}{A character vector.}

\item{.pattern}{Pattern to look for.}
}
\value{
An integer value.
}
\description{
Search each line in \code{.lines} for the regex \code{.pattern} and return
the first index of a match. Yields an error if the pattern cannot be located.
}
