% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_rows}
\alias{mutate_rows}
\title{Extension of \code{\link[dplyr]{mutate}} by a predicate function}
\usage{
mutate_rows(.df, .p, .ignore_na = FALSE, ...)
}
\arguments{
\item{.df}{A \code{\link[tibble]{tibble}}.}

\item{.p}{A predicate function.}

\item{.ignore_na}{Should NA values in the evaluation of the predicate
function be ignored? Defaults to \code{FALSE}.}

\item{...}{Name-value pairs passed to \code{\link[dplyr]{mutate}}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Use a predicate function \code{.p} to mutate only a subset of rows where the
predicate is satisfied. Creating new columns in combination with a predicate
creates by default \code{NA} values in rows where the predicate is not
satisfied.
}
\examples{
mutate_rows(
  tibble::tibble(
    x = c(-1, 0, -1),
    y = c(1, 1, 1)
  ),
  x < 0,
  y = y + 1,
  z = y + 1
)
}
