% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_tibble.R
\name{random_tibble}
\alias{random_tibble}
\title{Create a random \code{\link[tibble]{tibble}}}
\usage{
random_tibble(
  .col_names,
  .clean_col_names = TRUE,
  .col_types = NULL,
  .n_rows = 100
)
}
\arguments{
\item{.col_names}{Column names for the random \code{\link[tibble]{tibble}}.}

\item{.clean_col_names}{Should column names be cleaned? Defaults to
\code{TRUE}.}

\item{.col_types}{A list of column types. If \code{NULL} (the default),
then all columns will be of type \code{character}.}

\item{.n_rows}{Number of rows for the random \code{\link[tibble]{tibble}}.}
}
\value{
A randomly created \code{\link[tibble]{tibble}} with specified column
types.
}
\description{
Create a random \code{\link[tibble]{tibble}}
}
\examples{
random_tibble(
  c("x_1", "X 2"),
  .n_rows = 10
)
}
