% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sub}
\alias{read_sub}
\title{Subset lines and read into a \code{\link[tibble]{tibble}}}
\usage{
read_sub(.file, .pattern = NULL, .delim = ",", ...)
}
\arguments{
\item{.file}{A path to a file or a connection.}

\item{.pattern}{Pattern to look for.}

\item{.delim}{Single character used to separate fields within a record.}

\item{...}{Further arguments passed to \code{\link[readr]{read_delim}}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Read in the lines in \code{.file}, (optionally) identify the first line to
match \code{.pattern} and starting here read into a
\code{\link[tibble]{tibble}} with \code{.delim} as separator and clean names
via \code{\link[janitor]{clean_names}}. All lines
are read in if no \code{.pattern} is passed (the default).
}
\examples{
lines <- c(
  "key_1,val_1",
  "key_2,val_2",
  "",
  "y,x_1,x_2",
  "1,low,0.989",
  "0,medium,-1.923"
 )
 temp_file <- tempfile()
 writeLines(lines, temp_file)
 read_sub(temp_file, "(?=.*y)(?=.*x_1)(?=.*x_2)")

}
