% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{split_groups}
\alias{split_groups}
\title{Split by group and name by group key}
\usage{
split_groups(.df, .group_vars)
}
\arguments{
\item{.df}{A \code{\link[tibble]{tibble}}.}

\item{.group_vars}{The column names to group by.}
}
\value{
A list of \code{\link[tibble]{tibble}}s split by group named by keys.
}
\description{
This wrapper function performs a group split via
\code{\link[dplyr]{group_split}} and renames the resulting list elements
with the group keys.
}
\examples{
split_groups(
  datasets::iris,
  "Species"
)
}
